Split basic files
